refactor(app): move the four non-ContentView view types out of ContentView.swift (#1160 slice 6c)
Final slice of the ContentView decomposition. MainTerminalDetailView,
MainToolbarTitleBreadcrumb, WorkspaceConnectingOverlay, and
ContentViewPreviewHost were always their own top-level view types; only their
file was shared with ContentView. Each moves to its own file, unchanged.

ContentView.swift drops 295 lines (3,526 -> 3,231) with no identity risk at
all: moving a type between files in the same module is invisible to SwiftUI —
no view was rewrapped, no modifier chain changed, no .id touched.

Two things 6c was scoped to do are deliberately NOT here, with reasons:

The detail-content subview extraction turned out to be already done. All four
branches of detailContent (EmbeddedWebNextDetailView, WebSourceDetailView,
RepoLandingView, MainTerminalDetailView) are separate View types today; what
remained was file separation, which is what this slice does. Wrapping them in
another view would have added a ~40-parameter forwarding layer for no
invalidation benefit.

The deferred 6a modifier relocation is dropped. It was measured in #1194 to
cost lines rather than save them (~11 onChange keys plus ~14 closures that
cannot be bundled), it adds no testability, and it would collide directly with
#1194 and #1196, both open and approved against exactly those lines.

Two clusters the original plan listed as out of scope remain: the
open-in-editor focused-action family and the automation gesture-verb wiring.
Recorded as follow-ups in the closing summary rather than rushed in here.

Closes #1160
