+++ Sources/WorkspaceManager/Views/MainWindow/ContentView.swift
@@ -3229,298 +3229,3 @@ private final class ContextMenuActionItem: NSObject {
 }
 
 // MARK: - Main Terminal (Host-pinned)
-
-struct MainTerminalDetailView: View {
-    @ObservedObject var appCommandState: AppCommandState
-    let selectedWorkspace: Workspace?
-    let selectedRepo: Repo?
-    let activeHostSession: HostTerminalSession?
-    let hostTerminalSessions: [HostTerminalSession]
-    let visibleHostTerminalSessions: [HostTerminalSession]
-    let activeHostTerminalSessionID: UUID?
-    let activeTabTree: TileTreeState?
-    let resolveTileSession: (TileID) -> HostTerminalSession?
-    let resolveTileID: (HostTerminalSession) -> TileID
-    let hostSurfaceStore: SurfaceStore
-    let tabTitleOverrides: [UUID: String]
-    let agentStatuses: [AgentSessionStatus]
-    let terminalContextMenuProvider: (HostTerminalSession) -> NSMenu?
-    let onSetSplitRatio: (SplitID, CGFloat) -> Void
-    var onSelectTerminalTab: ((UUID) -> Void)?
-    var onCloseTerminalTab: ((UUID) -> Void)?
-    var onRenameTerminalTab: ((UUID, String?) -> Void)?
-    var onTerminalCloseConfirmationRequired: ((UUID) -> Void)?
-    var onTerminalProcessExit: ((UUID) -> Void)?
-    @Binding var selectedCodePreview: CodePreviewSelection?
-    @Binding var isTerminalPanelVisible: Bool
-    let onFileSelected: (CodePreviewSelection) -> Void
-    let availableEditors: [ExternalEditorDescriptor]
-    let defaultEditor: ExternalEditorDescriptor?
-    let onOpenInDefaultEditor: () -> Void
-    let onOpenInEditor: (ExternalEditorID) -> Void
-    let onCodePreviewSaved: () -> Void
-    /// Close the preview, routed through the dirty-navigation guard so unsaved edits prompt first.
-    let onCloseCodePreview: () -> Void
-    let rightPaneStateStore: RightPaneStateStore
-    @Binding var isRightPaneVisible: Bool
-
-    var body: some View {
-        HSplitView {
-            previewAndTerminalPanel
-                .frame(minWidth: 400)
-
-            // Collapsible right pane
-            if isRightPaneVisible {
-                if let selectedWorkspace {
-                    let state = rightPaneStateStore.state(for: selectedWorkspace)
-                    RightPaneView(
-                        workspace: selectedWorkspace,
-                        state: state,
-                        diagnosticWorkspaceDirectories: diagnosticWorkspaceDirectories,
-                        agentStatuses: agentStatuses,
-                        timelineHostSessionID: timelineHostSessionID(for: selectedWorkspace),
-                        onFileSelected: onFileSelected
-                    )
-                    .rightPaneWidth(for: state)
-                } else if let selectedRepo {
-                    let state = rightPaneStateStore.state(for: selectedRepo)
-                    RightPaneView(
-                        repo: selectedRepo,
-                        state: state,
-                        diagnosticWorkspaceDirectories: diagnosticWorkspaceDirectories,
-                        agentStatuses: agentStatuses,
-                        onFileSelected: onFileSelected
-                    )
-                    .rightPaneWidth(for: state)
-                }
-            }
-        }
-        .navigationTitle(navigationTitle)
-    }
-
-    private var diagnosticWorkspaceDirectories: [URL] {
-        var seen = Set<String>()
-        var directories: [URL] = []
-
-        // Split panes share their primary's directory, which is already in `hostTerminalSessions`, so
-        // the tab tree contributes no new diagnostic directories.
-        var candidateDirectories = hostTerminalSessions.map(\.directoryURL)
-        if let selectedWorkspaceDirectory = selectedWorkspace?.localDirectoryURL {
-            candidateDirectories.append(selectedWorkspaceDirectory)
-        }
-        if let selectedRepo {
-            candidateDirectories.append(selectedRepo.localURL)
-        }
-
-        for directory in candidateDirectories {
-            let path = directory.standardizedFileURL.resolvingSymlinksInPath().path
-            if seen.insert(path).inserted {
-                directories.append(URL(fileURLWithPath: path, isDirectory: true))
-            }
-        }
-
-        return directories
-    }
-
-    private func timelineHostSessionID(for workspace: Workspace) -> UUID? {
-        guard let workspaceDirectory = workspace.localDirectoryURL else { return nil }
-        let workspacePath = normalizedPath(workspaceDirectory)
-
-        if let activeHostTerminalSessionID,
-            let activeVisibleSession = visibleHostTerminalSessions.first(where: {
-                $0.id == activeHostTerminalSessionID && normalizedPath($0.directoryURL) == workspacePath
-            })
-        {
-            return activeVisibleSession.id
-        }
-
-        if let visibleSession = visibleHostTerminalSessions.first(where: {
-            normalizedPath($0.directoryURL) == workspacePath
-        }) {
-            return visibleSession.id
-        }
-
-        return hostTerminalSessions.first {
-            normalizedPath($0.directoryURL) == workspacePath
-        }?.id
-    }
-
-    private func normalizedPath(_ url: URL) -> String {
-        url.standardizedFileURL.resolvingSymlinksInPath().path
-    }
-
-    @ViewBuilder
-    private var previewAndTerminalPanel: some View {
-        VStack(spacing: 0) {
-            previewAndTerminalPanelContent
-        }
-    }
-
-    @ViewBuilder
-    private var previewAndTerminalPanelContent: some View {
-        if let selectedCodePreview {
-            if isTerminalPanelVisible {
-                VSplitView {
-                    CodeFilePreviewView(
-                        appCommandState: appCommandState,
-                        selection: selectedCodePreview,
-                        editorOptions: availableEditors,
-                        defaultEditor: defaultEditor,
-                        onOpenInDefaultEditor: onOpenInDefaultEditor,
-                        onOpenInEditor: onOpenInEditor,
-                        onSaved: onCodePreviewSaved,
-                        onClose: onCloseCodePreview
-                    )
-                    .frame(minHeight: 220)
-
-                    hostTerminalPanel
-                        .frame(minHeight: 160)
-                }
-            } else {
-                CodeFilePreviewView(
-                    appCommandState: appCommandState,
-                    selection: selectedCodePreview,
-                    editorOptions: availableEditors,
-                    defaultEditor: defaultEditor,
-                    onOpenInDefaultEditor: onOpenInDefaultEditor,
-                    onOpenInEditor: onOpenInEditor,
-                    onSaved: onCodePreviewSaved,
-                    onClose: onCloseCodePreview
-                )
-            }
-        } else {
-            hostTerminalPanel
-        }
-    }
-
-    private var hostTerminalPanel: some View {
-        HostTerminalSessionStack(
-            sessions: visibleHostTerminalSessions,
-            activeSessionID: activeHostTerminalSessionID,
-            tree: activeTabTree,
-            resolveSession: resolveTileSession,
-            resolveTileID: resolveTileID,
-            surfaceStore: hostSurfaceStore,
-            tabTitleOverrides: tabTitleOverrides,
-            onSetSplitRatio: onSetSplitRatio,
-            onSelectTab: onSelectTerminalTab,
-            onCloseTab: onCloseTerminalTab,
-            onRenameTab: onRenameTerminalTab,
-            onCloseConfirmationRequired: onTerminalCloseConfirmationRequired,
-            onTerminalProcessExit: onTerminalProcessExit,
-            contextMenuProvider: terminalContextMenuProvider
-        )
-    }
-
-    private var navigationTitle: String {
-        MainWindowPresentationController().toolbarTitle(
-            selectedWorkspace: selectedWorkspace,
-            selectedRepo: selectedWorkspace?.sourceRepo ?? selectedRepo,
-            activeHostSession: activeHostSession
-        )?.windowTitle ?? "WorkSpaces"
-    }
-}
-
-private struct MainToolbarTitleBreadcrumb: View {
-    let title: MainWindowToolbarTitle
-    let faviconSource: WebSource?
-    let onOpenRepoOverview: () -> Void
-    let onOpenRepoTerminal: () -> Void
-
-    var body: some View {
-        HStack(spacing: 5) {
-            Button(action: onOpenRepoOverview) {
-                repoIcon
-            }
-            .buttonStyle(.plain)
-            .help("Open Repo Overview")
-            .accessibilityLabel("Open Repo Overview")
-            .accessibilityIdentifier("main-toolbar.repo-overview")
-
-            Button(action: onOpenRepoTerminal) {
-                Text(title.repoName)
-                    .font(.callout.weight(.semibold))
-                    .lineLimit(1)
-            }
-            .buttonStyle(.plain)
-            .help("Open \(title.repoName) Terminal")
-            .accessibilityIdentifier("main-toolbar.repo-terminal")
-
-            if let workspaceName = title.workspaceName {
-                Text("/")
-                    .foregroundStyle(.secondary)
-                    .accessibilityHidden(true)
-
-                Text(workspaceName)
-                    .font(.callout)
-                    .lineLimit(1)
-                    .truncationMode(.tail)
-                    .accessibilityIdentifier("main-toolbar.workspace-title")
-            }
-        }
-        .frame(maxWidth: 360)
-        .accessibilityIdentifier("main-toolbar.title")
-    }
-
-    @ViewBuilder
-    private var repoIcon: some View {
-        if let faviconSource {
-            WebSourceFaviconView(source: faviconSource)
-        } else {
-            Image(systemName: "folder.fill")
-                .font(.system(size: 13, weight: .semibold))
-                .foregroundStyle(.secondary)
-                .frame(width: 16, height: 16)
-        }
-    }
-}
-
-struct WorkspaceConnectingOverlay: View {
-    let workspaceName: String?
-
-    var body: some View {
-        ZStack {
-            Color.black.opacity(0.5)
-            VStack(spacing: 12) {
-                ProgressView()
-                    .controlSize(.large)
-                Text("Connecting to \(workspaceName ?? "workspace")...")
-                    .font(.headline)
-                    .foregroundStyle(.white)
-            }
-            .padding(24)
-            .background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 12))
-        }
-    }
-}
-
-#Preview {
-    ContentViewPreviewHost()
-        .modelContainer(for: [Repo.self, Workspace.self, WebSource.self], inMemory: true)
-}
-
-private struct ContentViewPreviewHost: View {
-    @State private var deepLinkState = WorkspaceDeepLinkState()
-    @State private var lastSurfaceRawValue = ""
-    @StateObject private var appCommandState = AppCommandState()
-    @StateObject private var tileTreeStore = TileTreeStore()
-    @StateObject private var workspaceProviderSetupCoordinator = WorkspaceProviderSetupCoordinator()
-    @StateObject private var hostLumeSmokeAutomation = HostLumeSmokeAutomationController(
-        environment: [:]
-    )
-    @StateObject private var desktopUISmokeAutomation = DesktopUISmokeAutomationController(
-        environment: [:]
-    )
-
-    var body: some View {
-        ContentView(
-            deepLinkState: $deepLinkState,
-            lastSurfaceRawValue: $lastSurfaceRawValue,
-            appCommandState: appCommandState,
-            tileTreeStore: tileTreeStore,
-            workspaceProviderSetupCoordinator: workspaceProviderSetupCoordinator,
-            hostLumeSmokeAutomation: hostLumeSmokeAutomation,
-            desktopUISmokeAutomation: desktopUISmokeAutomation
-        )
-    }
-}
