+++ Sources/WorkspaceManager/Views/MainWindow/ContentViewPreviewHost.swift
@@ -1,0 +1,40 @@
+//
+//  ContentViewPreviewHost.swift
+//  WorkspaceManager
+//
+//  Owns the stores ContentView requires so the SwiftUI preview can construct one.
+//
+
+import SwiftUI
+import WorkspaceManagerCore
+
+#Preview {
+    ContentViewPreviewHost()
+        .modelContainer(for: [Repo.self, Workspace.self, WebSource.self], inMemory: true)
+}
+
+struct ContentViewPreviewHost: View {
+    @State private var deepLinkState = WorkspaceDeepLinkState()
+    @State private var lastSurfaceRawValue = ""
+    @StateObject private var appCommandState = AppCommandState()
+    @StateObject private var tileTreeStore = TileTreeStore()
+    @StateObject private var workspaceProviderSetupCoordinator = WorkspaceProviderSetupCoordinator()
+    @StateObject private var hostLumeSmokeAutomation = HostLumeSmokeAutomationController(
+        environment: [:]
+    )
+    @StateObject private var desktopUISmokeAutomation = DesktopUISmokeAutomationController(
+        environment: [:]
+    )
+
+    var body: some View {
+        ContentView(
+            deepLinkState: $deepLinkState,
+            lastSurfaceRawValue: $lastSurfaceRawValue,
+            appCommandState: appCommandState,
+            tileTreeStore: tileTreeStore,
+            workspaceProviderSetupCoordinator: workspaceProviderSetupCoordinator,
+            hostLumeSmokeAutomation: hostLumeSmokeAutomation,
+            desktopUISmokeAutomation: desktopUISmokeAutomation
+        )
+    }
+}
